COP-9551 - Code clean up for UI layout
